refactor: Use meaningful package names
